+++ src/main/java/com/java6/datn/DTO/OrderItemDTO.java
@@ -6,9 +6,9 @@ import java.math.BigDecimal;
 
 @Data
 public class OrderItemDTO {
-    private Long orderItemID;
-    private Long orderID;
-    private Long productID;
+    private Integer orderItemID;
+    private Integer orderID;
+    private Integer productID;
     private Integer quantity;
     private BigDecimal price;
 }
