+++ src/main/java/com/java6/datn/Repository/CartItemRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CartItemRepository extends JpaRepository<CartItem, Long> {
-    List<CartItem> findByUserUserID(Long userID);
+public interface CartItemRepository extends JpaRepository<CartItem,Integer> {
+    List<CartItem> findByUserUserID(Integer userID);
 }
