+++ src/main/java/com/java6/datn/Repository/ReviewRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> findByProductProductID(Long productID); // lấy danh sách review theo sản phẩm
-    List<Review> findByUserUserID(Long userID);           // lấy danh sách review của user
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
+    List<Review> findByProductProductID(Integer productID); // lấy danh sách review theo sản phẩm
+    List<Review> findByUserUserID(Integer userID);           // lấy danh sách review của user
 }
