+++ src/main/java/com/java6/datn/Service/CartItemService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface CartItemService {
     List<CartItemDTO> getAllCartItems();
-    List<CartItemDTO> getCartItemsByUser(Long userID);
-    CartItemDTO getCartItemById(Long id);
+    List<CartItemDTO> getCartItemsByUser(Integer userID);
+    CartItemDTO getCartItemById(Integer id);
     CartItemDTO createCartItem(CartItemDTO cartItemDTO);
-    CartItemDTO updateCartItem(Long id, CartItemDTO cartItemDTO);
-    void deleteCartItem(Long id);
+    CartItemDTO updateCartItem(Integer id, CartItemDTO cartItemDTO);
+    void deleteCartItem(Integer id);
 }
